git-svn-id: http://devel.cpl.upc.edu/recursos/svn/smeagol/trunk@1888 4ab63a03-c2d3-4037-8559-9a21b594da05

+++ lib/V2/Server/Controller/Resource.pm
@@ -41,13 +41,15 @@ sub default_GET {
 	my @message;
 
     if ($id) {
-		if(($module eq 'tag') && ($id_module)){
-		    $c->detach( 'get_relation_tag_resource', [$id, $id_module]);
-		}elsif(($module eq 'tag') && !($id_module)){
-			$c->response->location($c->uri_for('/tag')."/?resource=".$id);
-			#TODO: message: redireccio a la llista
-			$c->stash->{content}  = \@message;
-			$c->response->status(301); 
+		if((defined $module) && ($module eq 'tag')){
+			if($id_module){
+		    	$c->detach( 'get_relation_tag_resource', [$id, $id_module]);
+			}else{
+				$c->response->location($c->uri_for('/tag')."/?resource=".$id);
+				#TODO: message: redireccio a la llista
+				$c->stash->{content}  = \@message;
+				$c->response->status(301);
+			}
 		}else {
 	    	$c->detach( 'get_resource', [$id] );
 		}
@@ -124,7 +126,7 @@ sub default_POST {
     my $descr = $req->parameters->{description};
     my $info  = $req->parameters->{info};
     
-    if(($module eq 'tag') && ($id_module)){
+    if((defined $module) && ($module eq 'tag') && ($id_module)){
 		$c->detach( 'post_relation_tag_resource');
 	}
       
@@ -197,7 +199,7 @@ sub default_PUT {
     my ( $self, $c, $id, $module, $id_module ) = @_;
 
     if ($id) {
-		if(($module eq 'tag') && ($id_module)){
+		if((defined $module) && ($module eq 'tag') && ($id_module)){
 		    $c->forward( 'put_relation_tag_resource', [$id, $id_module]);
 		}else{
 		    $c->forward( 'put_resource', [$id] );
@@ -296,7 +298,7 @@ sub default_DELETE {
     my $req = $c->request;
     
 if ($id) {
-    if(($module eq 'tag') && ($id_module)){
+    if((defined $module) && ($module eq 'tag') && ($id_module)){
         $c->detach( 'delete_relation_tag_resource', [$id, $id_module]);
     }
     else {
